folder structure changed for making pip module
