+++ .gitignore
@@ -3,3 +3,4 @@
 *cache*
 build
 *extra*
+magurn.egg-info
