+++ pip_module_steps.md
@@ -1,0 +1,5 @@
+pip uninstall magurn
+y
+python setup.py sdist bdist_wheel
+pip install dist\magurn-1.0-py3-none-any.whl
+magurn
